backend/requirements.txt
anyio==3.6.2
APScheduler==3.6.3
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
beautifulsoup4==4.11.2
cachetools==4.2.2
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
cssbeautifier==1.14.7
defusedxml==0.8.0rc2
distlib==0.3.6
Django==3.2.16
django-debug-toolbar==2.2
django-filter==23.2
django-haystack==3.2.1
django-mail-templated==2.6.5
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djlint==1.19.16
djoser==2.1.0
EditorConfig==0.12.3
exceptiongroup==1.1.1
flake8==3.9.2
flake8-docstrings==1.6.0
h11==0.14.0
html-tag-names==0.1.2
html-void-elements==0.1.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==1.5.0
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jsbeautifier==1.14.7
MarkupSafe==2.1.3
mccabe==0.6.1
more-itertools==8.2.0
numpy==1.24.2
oauthlib==3.2.2
packaging==23.1
pandas==1.5.3
pandas-read-xml==0.3.1
pathspec==0.11.0
Pillow
pluggy==0.13.1
psycopg2-binary
py==1.11.0
pyarrow==11.0.0
pycodestyle==2.7.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==2.3.1
PyJWT==2.8.0
pyparsing==2.4.6
pysolr==3.9.0
pytest==6.2.5
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.7
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
regex==2022.10.31
requests==2.26.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.5.1
sorl-thumbnail==12.9.0
soupsieve==2.4
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
tornado==6.3.3
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.16
wcwidth==0.1.8
xmltodict==0.13.0
zipfile36==0.1.3
zipp==2.2.0